Move `std::io::IoSlice(Mut)` to `core::io`
